Talk to Folio Edge and handle response.
The URL is different from what I initially expected (thanks to @wwelling for the assistance).
Make the Port optional (to avoid explicitly requiring ":80" or ":443" in the URL).
Do the same for other parts of the path.

The configuration property names overlap with the folio xml property names.
Rename the configuration properties to start with "FIELD_" instead of "NODE_" and make them publically accessible from the factory.
The location of things could be improved, but I am saving that for a later time.

Use RestTemplate to perform the HTTP request to allow for better handling of HTTPS as well as HTTP redirects.
(Again, thanks to @wwelling.)

After reviewing the marcxml response from Folio Edge, I have discovered that many of the XML elements need "marc:" prefixed to their element names.
There is also are case-sensitivity concerns (`getRecord` is now `GetRecord`).
The code now requires a Prefix to be passed and an additional function is provided to default the prefix to an empty string for compatibility with the existing Voyager code.

Beyond just the names, the XML structure was slightly different than expected and the code has been updated to handle this.

Use static constants more often.