Use a more consistent and correct approach.
Due to a misunderstanding on my part, what I thought was "holdings" are actually (parts of) the "holding items".
Change the code accordingly to properly create the items with the information currently available.

The code is now more consistent with expected results but there is additional information that needs to be retrieved that is currently not available.
This only strives to utilize what is currently available with the expectation of future changes.