Issue 91: changes to fines implementation.
Use a patron controller instead of fines.
- Fix naming of the fines method.
- Avoid using `@WeaverUser`, instead passing the uin string.
- Change the path to `/patron/{uin}/fines`.

When a 404 is returned from the folio endpoint, RestTemplate throws a 404 Exception that does not get translated to a 404 to the client.
Instead a 500 is presented to the client.
Utilize `@ControllerAdvice` and handle these cases to ensure the 404 is correctly received by the client.

Bubble the `ParseException` instead of handling it in `folioDateToDate()`.
This required using an iterator to ensure that the `ParseException` is correctly bubbled.