Issue 91: implement tests and restdocs (stripping out much of weaver).
There is a lot of weaver used here that is causing significant problems with testing.
Much of the unused weaver functionality is stripped out.
Much of the code stripped out is from @wwelling.

This utilizes MockRestServiceServer.

In particular, this provides testing against return responses from, say the folio service (albeit mocked).
A general test is provided for the client side exceptions and a general test is provided for the service side exceptions.
Specific tests are provided for Voyager response and for the custom Not Found Exception that should be triggered according to the `@ControllerAdvice` annotation.