Improve patron controller tests and fix problems.
Further generalize the tests.

The same UIN is used everywhere, make it a static constant string.
There already are folio and voyager catalog strings, use them.
The "loan" and "fines" strings used in the context of endpoints are used all over the place, also make them static constant strings.

Document the Path and Query parameters.

The code was adding the specialized "loan" and "fines" endpoint strings to the get() call of mockMvc.perform().
It should not be used as such as those are specially reserved exclusively for Path Parameters.
Instead, concatenate the string.

Massively reduce the repeated code, further abstracting most of the file.

Clean up the structure a bit.
It can be very confusing having helper functions randomly in between test functions.
Move all of the helper functions to the bottom and make them all private.
Further rename such functions to avoid confusing them with tests.