Issue 95: FOLIO implementation of request holds.
Implemented `getNodeDateValue()` to retreived the node datetime value similar to how `getNodeValue()` is implemented.

The test implementation revealed several problems that required code changes beyond what is needed to get Issue 95 resolved.

Discovered that there several more likely Datetime formats, so a custom utility, `FolioDatetime`, is implemented utilizing the Apache Commons `DateUtils`.

The `restTemplate.postForObject()` is returning `null` responses during tests.
Replace it with `restTemplate.postForEntity()`, which is essentially `restTemplate.exchange()` with the HTTP Method automatically set to POST.
Furthermore, instead of throwing a NULL exception, return the pre-defined Internal Server Error message, with NULL in the logs (this makes the exception handling behavior a little bit more consistent).

The `TokenUtility` is causing tests to fail because the tests are expected responses and the responses are (correctly) being cached.
Provide a `clearToken()` method and a `clearAll()` method that can be used by the tests.
Implement the appropriate additional testing for these new functions.

Make significant changes in the testing process to handle the Okapi responses alongside non-Okapi responses.

Use `DefaultResponseCreator` instead of `ResponseCreator`.

The changes revealed problems in the `FolioCatalogServiceTest` class.
Replace all of the mockito mocks and injection with the `MockRestServiceServer` via the `@RestClientTest` annotation.
The `@RestClientTest` annotation handles the appropriate Spring Boot loading without loading all of Spring Boot (only loads what is necessary to mock the HTTP service).
This annotation, however, conflicts with the `@WebMvcTest` annotation and can only be used when `@WebMvcTest` is not.

The use of `@RestClientTest` and `MockRestServiceServer` exposed more problems with the `FolioCatalogService` class.
Specifically, `okapiLogin()` may still get null pointer exceptions, even when `restTemplate.postForEntity()` is used.

Be more consistent with the structure of the mocked responses, moving them within the response folder.