Issue 117: actually generate the gh-pages documentation.
The generation was not being performed, bring in asciidoctor-maven-plugin to perform this task.
It looks like "generated-docs" is a second step in the process, so "generated-snippets" must be used for the snippets.
It also seems that this must be explicitly specified.

According to the documentation, "Note the backend is html5 by default.".
The problem here is that the current version (1.6.0) has xml as the default.
For forward compatibility purposes, explicitly set to `html5` (as opposed to `html`).