Issue 125: Loans, FeeFines (Charges), and Holds should all include itemId as well as instanceId.
The `itemId` is required by clients and needs to be made available.

Organize the Domain Models to have all of the "id" fields at the top.
This should make them easier to find and is hopefully a little bit more organized.
The original approach was to closely match the legacy structure as possible, which led to different locations of where the ID fields are presented.

`instanceId` is being included for consistency.
That is, we should either consistently have `instanceId` or not have it at all.
If it is decided we do not want `instanceId` on the Domain Model, then it can be removed, likely with little harm.

I also noticed and corrected where `holdId` was used when `requestId` is more consistent and hopefully less confusing.