Actually, edge-patron does use holdId instead of requestId for this specific case.
The example found here shows "body", but this "body" is the response body not the request body.
- https://s3.amazonaws.com/foliodocs/api/mod-patron/patron.html#patron_account__id__hold__holdid__cancel_post

Instead, looking at the example raml file, one can see that it is indeed the `holdId` and not the `requestId` despite `requestId` being used as `holdId` in other projects or other parts of the project.
- https://github.com/folio-org/edge-patron/blob/083584e2be1a28178b07e48913a8cb778b1865d3/ramls/examples/hold-cancellation.json
- https://github.com/folio-org/edge-patron/blob/65370b9d5cef92f4d678092ee87cb95a1c2fe98d/src/main/java/org/folio/edge/patron/model/HoldCancellation.java#L18