Improve documentation, add API Specification link, and restructure test class.
Add missing parameter/response snippets.
Utilize bold and italics to better display present text as appropriate.
Provide slightly more detailed descriptions.

Restructure the Patron Controller tests to have a middle-man class that separates the statics and the helper functions.
This makes it easier to navigate to the controller tests themselves by isolating the tests.
This should make the tests even easier to read.