Issue 184: Upgrade Weaver-Webservice-Core (2.1.1-RC9).
The spring upgrade introduces problems with URL parameter substitution.
A fix is also applied to this.

The `isAssignableFrom()` doesn't work well in other projects.
For consistency, this project is also updated to use string equals on the class.
```
    return CatalogService.class.toString().equals(parameter.getParameterType().toString());
```

When seeing an error like:
```
---------------------------------------------------
Exception in thread "main" java.lang.AssertionError
	at jdk.compiler/com.sun.tools.javac.util.Assert.error(Assert.java:155)
	at jdk.compiler/com.sun.tools.javac.util.Assert.check(Assert.java:46)
...
```

Ensure the `maven-compiler-plugin` plugin is present and add `forceJavacCompilerUse`, setting it to TRUE.