Issue 202-203: Implement Access Token and Refresh Token support.
This creates basic structures for utilizing the Access Token (AT) and Refresh Token (RT).
A `ZonedDateTime` is used to represent the `expires` HTTP cookie property.

The new behavior is now exposed as configuration settings.
These configuration settings are named and structured in exactly the same manner as in `spring-module-core`, `mod-workflow`, and `mod-camunda` for consistency reasons.

The `tenant` default is normally `diku`, but the existing behavior in this project is to use `public`.
This preserves the existing default of `public` for compatibility reasons.

A new property `expireOffset` is provided as a way to trigger the token re-authentication process before a token might expire.
This is done to reduce the chances of the token expiring in-flight.
This behavior can be effectively disabled by setting `expireOffset` to a value of `0`.