Fix BibId error logging and provide API Response payload with proper HTTP status codes.
If an invalid ID is passed to the system log shows a rather unhelpful null pointer exception.
Improve the error logging by throwing an exception with an explicit message.

The `ApiResponse` always returns `HTTP 200`, even on errors, not founds, etc...
Utilize the exception handlers to return proper HTTP status codes while still preserving the existing behavior of an API Response.

Many programs that depend on this might be hard-coded to always expect an HTTP 200, even for errors.
Returning a proper HTTP status code might require these programs to be updated.

There are some exceptions already being handled that pass along a JSON response and status code already.
These are left alone (they already do not return an `ApiResponse`).

This only implements the basic functionality of this HTTP status code for specific cases.
With this approach, all of the exceptions could be handled this way and the controllers could potentially have reduced need for some conditional checks in them.