Add HoldingsRequestError to handle more problems and perform other minor fixes.
The DEV environment is failing with during an upgrade process:
```json
{"meta":{"status":"SUCCESS","action":null,"message":"Your request was successful","id":null},"payload":{"ArrayList":[]}}
```

This should actually be an error.

Change the code to properly handle this error, thus yielding the following:
```json
{"meta":{"status":"ERROR","action":null,"message":"Failed to get holdings for folio catalog due to: Error 'badArgument': Identifier has invalid structure.","id":null},"payload":{}}
```

This also addresses the situation where multiple errors might be returned.
These errors are concatenated by the exception.

Fix tabbing/spacing in some of the recently changed files.