Improve handling of remote server error responses.
Bad error messages where HTML is being returned with a mimetype of `application/json` while containing details from the remote server.
The HTTP status code being returned is also incorrect.
The 404 should not be returned to the user because the current path is valid and does work.
If the remote server request fails, then an Internal Server error should instead be presented.

Catch `HttpServerErrorException` and `HttpClientErrorException` exceptions when calling `restTemplate.getForObject()`.

Provide a new exception, called `RemoteServerError` that provides specific error handling and reporting.
The `RemoteServerError` will present a simplified error message to the user, in JSON.
A log will be written to that provides additional details, such as the URL.
The original response payload will be in the log message, which is likely to be HTML.

The following is an example log snippet:
```
2025-11-21 10:11:22.058 ERROR 25530 --- [nio-9000-exec-1] e.t.c.exception.RestExceptionHandler     : Failed to GET https://localhost/bad_url_test/oai?verb=GetRecord&met
  404 Not Found: "<html><EOL><EOL><head><title>404 Not Found</title></head><EOL><EOL><body><EOL><EOL><center><h1>404 Not Found</h1></center><EOL><EOL><hr><center>nginx</center><EOL><EOL></bo
edu.tamu.catalog.exception.RemoteServerError: Failed to access remote server for folio catalog due to 404 NOT_FOUND.
  at edu.tamu.catalog.service.FolioCatalogService.restGet(FolioCatalogService.java:1561)
  at edu.tamu.catalog.service.FolioCatalogService.requestHoldings(FolioCatalogService.java:703)
  at edu.tamu.catalog.service.FolioCatalogService.getHoldingsByBibId(FolioCatalogService.java:182)
  at edu.tamu.catalog.controller.CatalogAccessController.getHoldings(CatalogAccessController.java:34)
```